Splitting the src/util directory into src/base and src/util.
This also breaks the libutil directory into libbase and libutil. By splitting this directory and library in two it should easier to break the cyclic dependencies for base, expr and options. The build and dependency order is now:
  lib < base < options < expr < util
The files that have been moved are roughly cvc4_assert, bitvector, rational, integer, output, and tls. The exact list of moved files is:
- bitvector.h
- bitvector.i
- cvc4_assert.cpp
- cvc4_assert.h
- exception.cpp
- exception.h
- exception.i
- gmp_util.h
- integer_cln_imp.cpp
- integer_cln_imp.h
- integer_gmp_imp.cpp
- integer_gmp_imp.h
- integer.h.in
- integer.i
- output.cpp
- output.h
- rational_cln_imp.cpp
- rational_cln_imp.h
- rational_gmp_imp.cpp
- rational_gmp_imp.h
- rational.h.in
- rational.i
- tls.h.in

This commit also add the forward declaration files:
- lemma_input_channel_forward.h
- lemma_output_channel_forward.h